+++ examples/manual_stats_reporting.py
@@ -30,7 +30,7 @@ def _manual_report(name):
     try:
         yield
     except Exception as e:
-        events.request_failure.fire(
+        events.request.fire(
             request_type="manual",
             name=name,
             response_time=(time() - start_time) * 1000,
@@ -39,11 +39,12 @@ def _manual_report(name):
         )
         raise
     else:
-        events.request_success.fire(
+        events.request.fire(
             request_type="manual",
             name=name,
             response_time=(time() - start_time) * 1000,
             response_length=0,
+            exception=None,
         )