+++ locust/test/test_runners.py
@@ -311,11 +311,13 @@ class TestLocustRunner(LocustTestCase):
             class task_set(TaskSet):
                 @task
                 def my_task(self):
-                    self.user.environment.events.request_success.fire(
+                    self.user.environment.events.request.fire(
                         request_type="GET",
                         name="/test",
                         response_time=666,
                         response_length=1337,
+                        exception=None,
+                        context={},
                     )
                     sleep(2)
 
@@ -334,11 +336,13 @@ class TestLocustRunner(LocustTestCase):
             class task_set(TaskSet):
                 @task
                 def my_task(self):
-                    self.user.environment.events.request_success.fire(
+                    self.user.environment.events.request.fire(
                         request_type="GET",
                         name="/test",
                         response_time=666,
                         response_length=1337,
+                        exception=None,
+                        context={},
                     )
                     sleep(2)
 
@@ -447,11 +451,13 @@ class TestMasterWorkerRunners(LocustTestCase):
 
             @task
             def incr_stats(l):
-                l.environment.events.request_success.fire(
+                l.environment.events.request.fire(
                     request_type="GET",
                     name="/",
                     response_time=1337,
                     response_length=666,
+                    exception=None,
+                    context={},
                 )
 
         with mock.patch("locust.runners.WORKER_REPORT_INTERVAL", new=0.3):
@@ -1263,7 +1269,7 @@ class TestMasterRunner(LocustTestCase):
         self.assertEqual(2, exception["count"])
 
     def test_exception_is_caught(self):
-        """ Test that exceptions are stored, and execution continues """
+        """Test that exceptions are stored, and execution continues"""
 
         class MyTaskSet(TaskSet):
             def __init__(self, *a, **kw):
@@ -1302,7 +1308,7 @@ class TestMasterRunner(LocustTestCase):
         self.assertEqual(2, exception["count"])
 
     def test_master_reset_connection(self):
-        """ Test that connection will be reset when network issues found """
+        """Test that connection will be reset when network issues found"""
         with mock.patch("locust.runners.FALLBACK_INTERVAL", new=0.1):
             with mock.patch("locust.rpc.rpc.Server", mocked_rpc()) as server:
                 master = self.get_runner()